requirements.txt
streamlit
tensorflow>=2.13
numpy
pillow
requests
